move componenets
