+++ apps/equivalent-exchange/package.json
@@ -14,6 +14,7 @@
     "update-types-local": "npx supabase gen types --lang=typescript --local > src/utils/data-access/database.types.ts"
   },
   "dependencies": {
+    "@eq-ex/ui": "workspace:*",
     "@radix-ui/react-dropdown-menu": "^2.1.14",
     "@radix-ui/react-slot": "^1.1.0",
     "@supabase/auth-ui-react": "^0.4.7",
