+++ apps/equivalent-exchange/src/app/(auth)/password-reset/page.tsx
@@ -1,5 +1,5 @@
-import PasswordUpdate from "@PNN/components/password-update";
-import Logout from "@PNN/components/logout";
+import PasswordUpdate from "@eq-ex/app/components/password-update";
+import Logout from "@eq-ex/app/components/logout";
 
 export default function PasswordUpdatePage() {
   return (
