+++ apps/equivalent-exchange/src/app/[...id]/layout.tsx
@@ -1,10 +1,10 @@
-import Logout from "@PNN/components/logout";
-import { getUser } from "@PNN/utils/data-access/data-acess";
+import Logout from "@eq-ex/app/components/logout";
+import { getUser } from "@eq-ex/app/utils/data-access/data-acess";
 import { redirect } from "next/navigation";
 
 export default async function Layout({
   children,
-  params = Promise.resolve( { id: [] }),
+  params = Promise.resolve({ id: [] }),
 }: {
   children: React.ReactNode;
   params: Promise<{ id: string[] }>;
