+++ apps/equivalent-exchange/src/components/rewards-card/rewards-card-container.tsx
@@ -6,7 +6,7 @@ import {
   getMaxCount,
   getRewardsCard,
   getUser,
-} from "@PNN/utils/data-access/data-acess";
+} from "@eq-ex/app/utils/data-access/data-acess";
 import Image from "next/image";
 
 export default async function RewardsCardContainer({
@@ -18,14 +18,13 @@ export default async function RewardsCardContainer({
   const card = await getRewardsCard(cardId); //TODO I should do this at the top level and pass the card
   const maxPoints = await getMaxCount(card.organization_id); // same as above
   const canModify = await canModifyCard(user.id, card.organization_id); // same as above
-  const url = `https://${(await headers()).get("host")}/${card.organization_id}/${
-    card.id
-  }`;
+  const url = `https://${(await headers()).get("host")}/${card.organization_id}/${card.id
+    }`;
   const qrCode = await toDataURL(url, {
     type: "image/webp",
     color: { dark: "#000000FF", light: "#00000000" },
   });
-  
+
   return (
     <>
       {user.id !== card.user_id && (
