+++ apps/equivalent-exchange/src/components/rewards-card/rewards-card.tsx
@@ -1,16 +1,16 @@
 "use client";
 import { PropsWithChildren, useEffect, useState, useCallback } from "react";
 import { motion } from "framer-motion";
-import front from "@PNN/assests/front.svg";
-import frontDark from "@PNN/assests/front-dark.svg";
-import back from "@PNN/assests/back.svg";
-import backDark from "@PNN/assests/back-dark.svg";
-import { Tables } from "@PNN/utils/data-access/database.types";
+import front from "@eq-ex/app/assests/front.svg";
+import frontDark from "@eq-ex/app/assests/front-dark.svg";
+import back from "@eq-ex/app/assests/back.svg";
+import backDark from "@eq-ex/app/assests/back-dark.svg";
+import { Tables } from "@eq-ex/app/utils/data-access/database.types";
 import PunchNode from "./punch-node";
-import { Button } from "@PNN/components/ui/button";
-import { getStamps, redeemRewards } from "@PNN/utils/data-access/data-acess";
-import { Skeleton } from "../ui/skeleton";
-import { useSupabaseRealtimeSubscription } from "@PNN/hooks/supabase-real-time-subscription";
+import { Button } from "@eq-ex/ui";
+import { getStamps, redeemRewards } from "@eq-ex/app/utils/data-access/data-acess";
+import { Skeleton } from "@eq-ex/ui";
+import { useSupabaseRealtimeSubscription } from "@eq-ex/app/hooks/supabase-real-time-subscription";
 import confetti from "canvas-confetti";
 import { RealtimePostgresChangesPayload } from "@supabase/supabase-js";
 import { useTheme } from "next-themes";
