--- apps/equivalent-exchange/src/components/ui/skeleton.tsx
@@ -1,18 +1,0 @@
-import { cn } from "@PNN/utils/shadcn/utils";
-
-function Skeleton({
-  className,
-  ...props
-}: React.HTMLAttributes<HTMLDivElement>) {
-  return (
-    <div
-      className={cn(
-        "animate-pulse rounded-md bg-stone-100 dark:bg-stone-800",
-        className
-      )}
-      {...props}
-    />
-  );
-}
-
-export { Skeleton };
