+++ packages/ui/README.md
@@ -1,0 +1,50 @@
+# @equivalent-exchange/ui
+
+A shared UI component library for the Equivalent Exchange workspace.
+
+## Features
+
+- TypeScript support
+- Tailwind CSS styling
+- Radix UI primitives
+- Built for Next.js applications
+- Hot reloading during development
+
+## Development
+
+```bash
+# Install dependencies
+pnpm install
+
+# Development mode (watch for changes)
+pnpm dev
+
+# Build the library
+pnpm build
+
+# Run linting
+pnpm lint
+
+# Type checking
+pnpm type-check
+```
+
+## Usage
+
+```typescript
+import { ComponentName } from '@equivalent-exchange/ui'
+
+// Use components in your app
+```
+
+## Adding Components
+
+1. Create your component in `src/components/`
+2. Export it from `src/components/index.ts`
+3. The component will be automatically available for import
+
+## Adding Utilities
+
+1. Create your utility in `src/utils/`
+2. Export it from `src/utils/index.ts`
+3. The utility will be automatically available for import
