+++ packages/ui/package.json
@@ -1,0 +1,43 @@
+{
+    "name": "@eq-ex/ui",
+    "version": "0.1.0",
+    "private": true,
+    "main": "./src/index.ts",
+    "types": "./src/index.ts",
+    "exports": {
+        ".": "./src/index.ts",
+        "./components": "./src/components/index.ts",
+        "./utils": "./src/utils/index.ts"
+    },
+    "scripts": {
+        "build": "tsc",
+        "dev": "tsc --watch",
+        "lint": "eslint src/**/*.{ts,tsx}",
+        "type-check": "tsc --noEmit"
+    },
+    "peerDependencies": {
+        "react": "^19.1.0",
+        "react-dom": "^19.1.0"
+    },
+    "dependencies": {
+        "@radix-ui/react-dropdown-menu": "^2.1.14",
+        "@radix-ui/react-slot": "^1.1.0",
+        "class-variance-authority": "^0.7.1",
+        "clsx": "^2.1.1",
+        "embla-carousel-react": "^8.5.1",
+        "framer-motion": "^12.9.0",
+        "lucide-react": "^0.331.0",
+        "next-themes": "^0.4.6",
+        "tailwind-merge": "^2.5.5",
+        "tailwindcss-animate": "^1.0.7"
+    },
+    "devDependencies": {
+        "@types/react": "^19.1.0",
+        "@types/react-dom": "^19.1.0",
+        "autoprefixer": "10.4.14",
+        "eslint": "8.41.0",
+        "postcss": "^8.4.49",
+        "tailwindcss": "3.3.2",
+        "typescript": "5.1.3"
+    }
+}
