+++ packages/ui/src/components/theme-toggle.tsx
@@ -11,13 +11,17 @@ import {
     DropdownMenuTrigger,
 } from "./dropdown-menu"
 
-export function ModeToggle() {
+export interface ModeToggleProps {
+    className?: string
+}
+
+export function ModeToggle({ className }: ModeToggleProps) {
     const { setTheme } = useTheme()
 
     return (
         <DropdownMenu>
             <DropdownMenuTrigger asChild>
-                <Button variant="outline" size="icon">
+                <Button variant="outline" size="icon" className={className}>
                     <Sun className="h-[1.2rem] w-[1.2rem] rotate-0 scale-100 transition-all dark:-rotate-90 dark:scale-0" />
                     <Moon className="absolute h-[1.2rem] w-[1.2rem] rotate-90 scale-0 transition-all dark:rotate-0 dark:scale-100" />
                     <span className="sr-only">Toggle theme</span>
