+++ packages/ui/src/utils/index.ts
@@ -1,0 +1,2 @@
+// Export all utilities here
+export { cn } from './cn'
